+++ signing/src/main/java/tech/pegasys/web3signer/signing/InMemoryValidatorManager.java
@@ -12,22 +12,13 @@
  */
 package tech.pegasys.web3signer.signing;
 
-import tech.pegasys.teku.bls.BLSKeyPair;
-import tech.pegasys.teku.bls.BLSSecretKey;
-import tech.pegasys.web3signer.bls.keystore.KeyStore;
-import tech.pegasys.web3signer.bls.keystore.KeyStoreValidationException;
-import tech.pegasys.web3signer.bls.keystore.model.KeyStoreData;
-import tech.pegasys.web3signer.signing.config.metadata.SignerOrigin;
-
 import java.util.Locale;
 import java.util.concurrent.ExecutionException;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.apache.tuweni.bytes.Bytes;
-import org.apache.tuweni.bytes.Bytes32;
 
 /**
  * In-memory only validator manager that skips writing keystores to disk. Validators imported via
@@ -46,6 +37,11 @@ public class InMemoryValidatorManager implements ValidatorManager {
   }
 
   @Override
+  public ObjectMapper getJsonMapper() {
+    return objectMapper;
+  }
+
+  @Override
   public void deleteValidator(final Bytes publicKey) {
     try {
       signerProvider.removeSigner(publicKey.toHexString()).get();
@@ -58,31 +54,20 @@ public class InMemoryValidatorManager implements ValidatorManager {
   }
 
   @Override
-  public void addValidator(final Bytes publicKey, final String keystore, final String password) {
+  public void addValidator(final BlsArtifactSigner signer) {
     try {
-      final BlsArtifactSigner signer = decryptKeystoreAndCreateSigner(keystore, password);
-
       signerProvider.addSigner(signer).get();
-
       LOG.info(
           "Validator added to memory only (no files written): {}",
           signer.getIdentifier().toLowerCase(Locale.ROOT));
-    } catch (InterruptedException | ExecutionException | KeyStoreValidationException e) {
+    } catch (InterruptedException | ExecutionException e) {
       throw new IllegalStateException("Unable to add validator to memory", e);
     }
   }
 
-  private BlsArtifactSigner decryptKeystoreAndCreateSigner(
-      final String jsonKeystoreData, final String password) throws KeyStoreValidationException {
-    try {
-      final KeyStoreData keyStoreData =
-          objectMapper.readValue(jsonKeystoreData, KeyStoreData.class);
-      final Bytes privateKey = KeyStore.decrypt(password, keyStoreData);
-      final BLSKeyPair keyPair = new BLSKeyPair(BLSSecretKey.fromBytes(Bytes32.wrap(privateKey)));
-
-      return new BlsArtifactSigner(keyPair, SignerOrigin.FILE_KEYSTORE);
-    } catch (final JsonProcessingException e) {
-      throw new KeyStoreValidationException("Failed to parse keystore JSON", e);
-    }
+  @Override
+  public void postAddValidator(
+      final BlsArtifactSigner signer, final String jsonKeystoreData, final String password) {
+    // do nothing
   }
 }