+++ signing/src/main/java/tech/pegasys/web3signer/signing/ValidatorManager.java
@@ -12,11 +12,38 @@
  */
 package tech.pegasys.web3signer.signing;
 
+import tech.pegasys.teku.bls.BLSKeyPair;
+import tech.pegasys.teku.bls.BLSSecretKey;
+import tech.pegasys.web3signer.bls.keystore.KeyStore;
+import tech.pegasys.web3signer.bls.keystore.KeyStoreValidationException;
+import tech.pegasys.web3signer.bls.keystore.model.KeyStoreData;
+import tech.pegasys.web3signer.signing.config.metadata.SignerOrigin;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.tuweni.bytes.Bytes;
+import org.apache.tuweni.bytes.Bytes32;
 
 public interface ValidatorManager {
+  ObjectMapper getJsonMapper();
 
   void deleteValidator(final Bytes publicKey);
 
-  void addValidator(final Bytes publicKey, final String keystore, final String password);
+  void addValidator(final BlsArtifactSigner signer);
+
+  void postAddValidator(
+      final BlsArtifactSigner signer, final String jsonKeystoreData, final String password);
+
+  default BlsArtifactSigner decryptKeystore(final String jsonKeystoreData, final String password) {
+    try {
+      final KeyStoreData keyStoreData =
+          getJsonMapper().readValue(jsonKeystoreData, KeyStoreData.class);
+      final Bytes privateKey = KeyStore.decrypt(password, keyStoreData);
+      final BLSKeyPair keyPair = new BLSKeyPair(BLSSecretKey.fromBytes(Bytes32.wrap(privateKey)));
+
+      return new BlsArtifactSigner(keyPair, SignerOrigin.FILE_KEYSTORE);
+    } catch (final JsonProcessingException e) {
+      throw new KeyStoreValidationException("Failed to parse keystore JSON", e);
+    }
+  }
 }